+++ CHANGELOG.md
@@ -7,6 +7,11 @@ but small API changes may happen between MINOR versions.
 
 This document mainly describes API changes important to users of this library.
 
+## 0.9 - 2017-07-16
+
+* Add `\Rokka\Client\UrlHelper::addOptionsToUriString(string $url, $options)` and `\Rokka\Client\UrlHelper::addOptionsToUri(UriInterface $url, $options)` for easily adding stack options to an existing URL.
+* Drop support for PHP 5.5.  
+
 ## 0.8.0 - 2017-06-15
 
 * Add overwrite parameter to `Rokka\Client\Image::createStack($stackName, $stackOperations, $organization = '',  $stackOptions = [], $overwrite = false)`. If set to true, a stack will be overwritten, if it already exists.